requirements.txt
# Core dependencies
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.24.0

# File handling and utilities
tqdm>=4.65.0
rich>=13.0.0

# Optional: GPU support (uncomment if using CUDA)
# torch>=2.0.0+cu118 -f https://download.pytorch.org/whl/torch_stable.html
# torchaudio>=2.0.0+cu118 -f https://download.pytorch.org/whl/torch_stable.html

# YouTube download support
yt-dlp>=2023.12.30

# MCP Server support
mcp>=1.0.0

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0 